==> Fluxo 1/02_Missing Analysis of all Indexes.ipynb <==
## This notebook analyses missingness of all indexes based on the dates of the base index S&P500

import pandas as pd
import numpy as np
from datetime import date, datetime
import holidays

business_dates = pd.bdate_range(start='2000-01-03', end='2020-03-18').date
business_dates = business_dates.tolist()

closed_market_dates = list()
years = list(range(2000,2021))
us_holidays = holidays.UnitedStates(years = years, expand = False, observed=False)
us_holidays[datetime(2019,4,19)] = 'Good Friday'
us_holidays[datetime(2018,12,5)] = 'Mourning Bush'
us_holidays[datetime(2018,3,30)] = 'Good Friday'
us_holidays[datetime(2017,4,14)] = 'Good Friday'
us_holidays[datetime(2017,1,2)] = "New Year's"
us_holidays[datetime(2016,3,25)] = 'Good Friday'
us_holidays[datetime(2016,12,26)] = 'Christmas'
us_holidays[datetime(2015,7,3)] = 'Independence Day'
us_holidays[datetime(2015,4,3)] = 'Good Friday'
us_holidays[datetime(2019,4,19)] = 'Good Friday'
us_holidays[datetime(2014,4,18)] = 'Good Friday'
us_holidays[datetime(2013,3,29)] = 'Good Friday'
us_holidays[datetime(2012,10,30)] = 'Hurricane Sandy'
us_holidays[datetime(2012,10,29)] = 'Hurricane Sandy'
us_holidays[datetime(2012,4,6)] = 'Good Friday'
us_holidays[datetime(2012,1,2)] = "New Year's"
us_holidays[datetime(2011,12,26)] = 'Christmas'
us_holidays[datetime(2011,4,22)] = 'Good Friday'
us_holidays[datetime(2010,12,24)] = 'Christmas'
us_holidays[datetime(2010,7,5)] = 'Independence Day'
us_holidays[datetime(2010,4,2)] = 'Good Friday'
us_holidays[datetime(2009,7,3)] = 'Independence Day'
us_holidays[datetime(2009,4,10)] = 'Good Friday'
us_holidays[datetime(2008,3,21)] = 'Good Friday'
us_holidays[datetime(2007,4,6)] = 'Good Friday'
us_holidays[datetime(2007,1,2)] = 'Mourning Ford'
us_holidays[datetime(2006,4,14)] = 'Good Friday'
us_holidays[datetime(2006,1,2)] = "New Year's"
us_holidays[datetime(2005,12,26)] = 'Christmas'
us_holidays[datetime(2005,3,25)] = 'Good Friday'
us_holidays[datetime(2004,12,24)] = 'Christmas'
us_holidays[datetime(2004,7,5)] = 'Independence Day'
us_holidays[datetime(2004,6,11)] = 'Mourning Reagan'
us_holidays[datetime(2004,4,9)] = 'Good Friday'
us_holidays[datetime(2003,4,18)] = 'Good Friday'
us_holidays[datetime(2002,3,29)] = 'Good Friday'
us_holidays[datetime(2001,9,14)] = '9-11'
us_holidays[datetime(2001,9,13)] = '9-11'
us_holidays[datetime(2001,9,12)] = '9-11'
us_holidays[datetime(2001,9,11)] = '9-11'
us_holidays[datetime(2001,4,13)] = 'Good Friday'
us_holidays[datetime(2000,4,21)] = 'Good Friday'

for date in business_dates:
    if(date not in list(us_holidays.keys())):
        closed_market_dates.append(date)
    else:
        print("Closed", date) 

def check_for_missingness(df, closed_market_dates, daily=True):
    at_least_one_value_missing = df.apply(lambda x: (len(x)-x.count())>=1, axis=1)
    out = at_least_one_value_missing.value_counts()
    if(out[False]==len(df)):
        print('There are no missing values on the original data.')
    else:
        print('There are ',np.absolute(out[False]-len(df)), ' missing values.')
    if(daily):    
        count = 0
        for date in closed_market_dates:
            if date.strftime('%Y-%m-%d') not in df['Date'].unique():
                count += 1
        print("There are", count, "dates missing from dataframe.")

v_sp500 = pd.read_csv('S&P500.csv', parse_dates=True)

check_for_missingness(v_sp500, closed_market_dates)

v_vix = pd.read_csv('VIX.csv', parse_dates=True)

check_for_missingness(v_vix, closed_market_dates)

v_dji = pd.read_csv('^DJI.csv', parse_dates=True)

check_for_missingness(v_dji, closed_market_dates)

v_ndx = pd.read_csv('^NDX.csv', parse_dates=True)

check_for_missingness(v_ndx, closed_market_dates)

v_n225 = pd.read_csv('N225.csv', parse_dates=True)

check_for_missingness(v_n225, closed_market_dates)

v_ftse = pd.read_csv('CompleteFTSE.csv', parse_dates=True)

check_for_missingness(v_ftse, closed_market_dates)

v_hsi = pd.read_csv('^HSI.csv', parse_dates=True)

check_for_missingness(v_hsi, closed_market_dates)

v_n100 = pd.read_csv('^N100.csv', parse_dates=True)

check_for_missingness(v_n100, closed_market_dates)

v_monthly_EMV_index =  pd.read_excel('EMV_Data.xlsx') 
v_monthly_EMV_index = v_monthly_EMV_index[:-1]

check_for_missingness(v_monthly_EMV_index, closed_market_dates, False)

v_infectious_EMV_index =  pd.read_csv('All_Infectious_EMV_Data.csv') 
v_infectious_EMV_index = v_infectious_EMV_index[:-1]

check_for_missingness(v_infectious_EMV_index, closed_market_dates, False)

v_gpr_index =  pd.read_excel('gpr_web_latest.xlsx') 
v_gpr_index = v_gpr_index[:-1]

check_for_missingness(v_gpr_index, closed_market_dates, False)

v_trade_uncertainty_index =  pd.read_excel('Trade_Uncertainty_Data.xlsx') 
v_trade_uncertainty_index = v_trade_uncertainty_index[:-1]

check_for_missingness(v_trade_uncertainty_index, closed_market_dates, False)
