==> Fluxo 1/05.01_Mean Imputation.ipynb <==
## This notebook inputs the missing values using Mean Imputation

import pandas as pd
from datetime import datetime
import numpy as np
from sklearn.impute import SimpleImputer

#### Choose wether the imputation should be done on the data set with daily or weekly granularity

#INPUT, OUTPUT = 'CompleteIndexes.csv','MeanImputed.csv'
INPUT, OUTPUT = 'CompleteWeeklyIndexes.csv', 'MeanWeeklyImputed.csv'

df = pd.read_csv(INPUT, parse_dates=True, index_col=0)
df_test = df.loc['2015-01-02':] #Selecting test data
df = df.loc['2000-01-01':'2015-01-01'] #Selecting training data
df.head()

df.tail()

imp = SimpleImputer(missing_values=np.nan, strategy='mean')
imp.fit(df)

df_imputed = imp.transform(df)
df_test_imputed = imp.transform(df_test)

df_final = pd.DataFrame(data=df_imputed, index=df.index, columns=df.columns)
df_test_final = pd.DataFrame(data=df_test_imputed, index=df_test.index, columns=df_test.columns)

df_final.head()

output = OUTPUT.split('.')
df_final.to_csv(OUTPUT, index = True)
df_test_final.to_csv(output[0]+ '_test.' + output[1], index = True)
