==> Fluxo 1/05.04_Regression Imputation.ipynb <==
## This notebook inputs the missing values using Regression

import pandas as pd
from datetime import datetime
import numpy as np
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from sklearn import preprocessing

#### Choose wether the imputation should be done on the data set with daily or weekly granularity

#INPUT, OUTPUT = 'CompleteIndexes.csv','RegressionImputed.csv'
INPUT, OUTPUT = 'CompleteWeeklyIndexes.csv', 'RegressionWeeklyImputed.csv'

df = pd.read_csv(INPUT, parse_dates=True, index_col=0)
df_test = df.loc['2015-01-02':] #Selecting test data
df = df.loc['2000-01-01':'2015-01-01'] #Selecting training data
df.head()

#Scaled data has zero mean and unit variance
scaler = preprocessing.StandardScaler()
scaler.fit(df)
array = scaler.transform(df)
array_test = scaler.transform(df_test)
df_scaled = pd.DataFrame(data=array, index=df.index, columns=df.columns)
df_test_scaled = pd.DataFrame(data=array_test, index=df_test.index, columns=df_test.columns)
df_scaled.head()

imp = IterativeImputer(max_iter=10, random_state=0)
imp.fit(df_scaled)
df_imputed = imp.transform(df_scaled)
df_test_imputed = imp.transform(df_test_scaled)
df_final = pd.DataFrame(data=df_imputed, index=df.index, columns=df.columns)
df_test_final = pd.DataFrame(data=df_test_imputed, index=df_test.index, columns=df_test.columns)

# Descaling the data
array_2 = scaler.inverse_transform(df_final)
df_final2 = pd.DataFrame(data=array_2, index=df.index, columns=df.columns)

array_test_2 = scaler.inverse_transform(df_test_final)
df_test_final2 = pd.DataFrame(data=array_test_2, index=df_test.index, columns=df_test.columns)

df_final2.head()

output = OUTPUT.split('.')
df_final2.to_csv(OUTPUT, index = True)
df_test_final2.to_csv(output[0]+ '_test.' + output[1], index = True)
